admin端对码功能修改

+++ jmp/mp-common/src/main/java/com/vpu/mp/service/pojo/shop/medical/goods/param/ExternalMatchedGoodsParam.java
@@ -20,7 +20,7 @@ public class ExternalMatchedGoodsParam {
     private String goodsCommonName;
     private String goodsAliasName;
     private String goodsQualityRatio;
-    private Integer source;
+    private Byte source;
     private Byte isMedical;
     private Byte isRx;
     private Byte medicalType;
